* App component moved to component folder
