+++ src/index.tsx
@@ -2,7 +2,7 @@ import { createRoot } from 'react-dom/client';
 import { StrictMode } from 'react';
 
 import './styles/index.scss';
-import { App } from 'src/app';
+import { App } from 'src/components/app';
 
 const domNode = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(domNode);
